Dataframe sorting sorted, as required.
The meta_data_df is now being sorted as rqeuired in #29. This was quite trickey to implement because the mix of letters, numbers and dashes meant that the sort methods wouldn't sort as expected. The index was first extracted and renamed. Then the indicator name e.g. 1-2-1 had to be split up into their own columns and sorted individually. Then the new columns were dropped and the dataframe returned