Making code for #44 and #45 work again
There were some bizzare changes in the code which stopped it working which I have now solved. The def of `get_name_list` had dissapeared, and under the code for ticket 45, the run was failing because the wrong column names were being imported and dropped. Now fixed. Additionally I have fixed the code for 44 so it now imports both `column_name` and `column_value` from the manually updated spreadsheet. Later the `sdmx_code` is mapped to `Value` and `column_value` to `Text`. Both are required for ticket #44.